mod exit_with_json
